Refactor to separate handle-related code
