+++ SettingsWatchdog/SettingsWatchdog.vcxproj
@@ -93,6 +93,7 @@
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="errors.cpp" />
+    <ClCompile Include="handles.cpp" />
     <ClCompile Include="SettingsWatchdog.cpp" />
     <ClCompile Include="logging.cpp" />
   </ItemGroup>
@@ -109,6 +110,7 @@
   <ItemGroup>
     <ClInclude Include="errors.hpp" />
     <ClInclude Include="logging.hpp" />
+    <ClInclude Include="handles.hpp" />
   </ItemGroup>
 
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
